Move logging to datamegh.api
